refactor(phase-5a): extract 5 boot modules out of desktop.ts
First wave of the boot decomposition. Pulls 368 LOC of pure setup
helpers out of desktop.ts (3,695 → 3,327) into focused
single-purpose modules under src/boot/. desktop.ts now imports
each helper rather than owning it; behavior unchanged. The
extracted helpers are smaller, individually testable, and the new
locations match the architecture-0.8.1 layout map in
docs/architecture.md.

- src/boot/origin.ts (20 LOC) — INITIAL_ORIGIN snapshot at module
  load. Lives in its own module so every boot consumer reaches
  the same captured value.
- src/boot/geometry.ts (82 LOC) — clampGeometryToViewport() +
  findDockEntryForUrl() + the VIEWPORT_CLAMP_MARGIN constant. Pure
  functions called by session restore.
- src/boot/session.ts (139 LOC) — restoreSession() rebuilds
  windows from the saved snapshot; openCurrentPage() handles the
  "no saved session" path.
- src/boot/shell-lifecycle.ts (71 LOC) — bindShellLifecycle()
  wires browser resize + visibilitychange to the SHELL_*
  hooks; wireSessionEvents() ties window-lifecycle CustomEvents
  to the saver.
- src/boot/menu-refresh.ts (218 LOC) — bindMenuRefresh() wires
  the live plugin-changed pipeline + the explicit refresh probe.
  Argument list collected into a single MenuRefreshDeps options
  object so future syncers don't grow the parameter list.

Phase 5 continues with the heavier extractions (createSessionSaver,
trackedFetch, bindTopWindowLinkInterceptor, the wp.desktop.* facade
assembly) in subsequent commits — those have richer dependency
shapes and need their own attention.

All green: lint, tsc --noEmit, vitest (106 files / 936 tests).
